# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Torrent streaming server.",
        "title": "Swagger Torrserver API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "{version.Version}"
    },
    "basePath": "/",
    "paths": {
...
        "/playlist": {
            "get": {
                "description": "Get HTTP link of torrent in M3U list.",
                "produces": [
                    "audio/x-mpegurl"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Get HTTP link of torrent in M3U list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Torrent hash",
                        "name": "hash",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "From last play file",
                        "name": "fromlast",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start the playlist at this file id (overrides fromlast)",
                        "name": "index",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
        "/stream": {
            "get": {
                "description": "Multi usage endpoint.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Multi usage endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Magnet/hash/link to torrent",
                        "name": "link",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File index in torrent",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Should preload torrent",
                        "name": "preload",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get statistics from torrent",
                        "name": "stat",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Should save torrent",
                        "name": "save",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get torrent as M3U playlist",
                        "name": "m3u",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get M3U from last played file",
                        "name": "fromlast",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start stream torrent",
                        "name": "play",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Set title of torrent",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Set poster link of torrent",
                        "name": "poster",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Set category of torrent, used in web: movie, tv, music, other",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data returned according to query"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}